refactor: Move "recent" api/queries to `insights/core` folder
